# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, file, plugin, mcp, skill. The config scope also
    grants bot and provider access.
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/file:
    post:
      tags: [Open API]
      summary: Upload a file
      operationId: uploadOpenApiFile
      x-astrbot-scope: file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    get:
      tags: [Open API]
      summary: Download an uploaded file by attachment ID
      operationId: downloadOpenApiFile
      x-astrbot-scope: file
      parameters:
        - $ref: "#/components/parameters/AttachmentIdQuery"
      responses:
        "200":
          description: File content or an error envelope
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
  /api/v1/plugins:
    get:
      tags: [Plugins]
      summary: List installed plugins
      operationId: listPlugins
      x-astrbot-scope: plugin
      parameters:
        - name: include_reserved
          in: query
          schema:
            type: boolean
            default: true
        - name: enabled
          in: query
          schema:
            type: boolean
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/by-id:
    get:
      tags: [Plugins]
      summary: Get plugin details by query ID
      operationId: getPluginById
      x-astrbot-scope: plugin
      parameters:
        - name: plugin_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Plugins]
      summary: Uninstall a plugin by query ID
      operationId: uninstallPluginById
      x-astrbot-scope: plugin
      parameters:
        - name: plugin_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                delete_config:
                  type: boolean
                delete_data:
                  type: boolean
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/config:
    get:
      tags: [Plugins]
      summary: Get plugin configuration by query ID
      operationId: getPluginConfigById
      x-astrbot-scope: plugin
      parameters:
        - name: plugin_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Plugins]
      summary: Save plugin configuration by body ID
      operationId: updatePluginConfigById
      x-astrbot-scope: plugin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plugin_id, config]
              properties:
                plugin_id:
                  type: string
                config:
                  $ref: "#/components/schemas/DynamicConfig"
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/config/schema:
    get:
      tags: [Plugins]
      summary: Get plugin configuration schema by query ID
      operationId: getPluginConfigSchemaById
      x-astrbot-scope: plugin
      parameters:
        - name: plugin_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/config-files:
    get:
      tags: [Plugins]
      summary: List uploaded files for a plugin file configuration item by query ID
      operationId: listPluginConfigFilesById
      x-astrbot-scope: plugin
      parameters:
        - name: plugin_id
          in: query
          required: true
          schema:
            type: string
        - name: config_key
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Plugins]
      summary: Upload files for a plugin file configuration item by query ID
      operationId: uploadPluginConfigFilesById
      x-astrbot-scope: plugin
      parameters:
        - name: plugin_id
          in: query
          required: true
          schema:
            type: string
        - name: config_key
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Plugins]
      summary: Delete a plugin file configuration upload by query ID
      operationId: deletePluginConfigFileById
      x-astrbot-scope: plugin
      parameters:
        - name: plugin_id
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginConfigFileDeleteRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/readme:
    get:
      tags: [Plugins]
      summary: Get plugin README content by query ID
      operationId: getPluginReadmeById
      x-astrbot-scope: plugin
      parameters:
        - name: plugin_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Text"
  /api/v1/plugins/changelog:
    get:
      tags: [Plugins]
      summary: Get plugin CHANGELOG content by query ID
      operationId: getPluginChangelogById
      x-astrbot-scope: plugin
      parameters:
        - name: plugin_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Text"
  /api/v1/plugins/reload:
    post:
      tags: [Plugins]
      summary: Reload a plugin by body ID
      operationId: reloadPluginById
      x-astrbot-scope: plugin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plugin_id]
              properties:
                plugin_id:
                  type: string
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/enabled:
    patch:
      tags: [Plugins]
      summary: Enable or disable a plugin by body ID
      operationId: setPluginEnabledById
      x-astrbot-scope: plugin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plugin_id, enabled]
              properties:
                plugin_id:
                  type: string
                enabled:
                  type: boolean
              additionalProperties: false
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/{plugin_id}:
    get:
      tags: [Plugins]
      summary: Get plugin details
      operationId: getPlugin
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                delete_config:
                  type: boolean
                delete_data:
                  type: boolean
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Plugins]
      summary: Uninstall a plugin
      operationId: uninstallPlugin
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                delete_config:
                  type: boolean
                delete_data:
                  type: boolean
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/{plugin_id}/config:
    get:
      tags: [Plugins]
      summary: Get plugin configuration
      operationId: getPluginConfig
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    put:
      tags: [Plugins]
      summary: Save plugin configuration
      operationId: updatePluginConfig
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DynamicConfig"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/{plugin_id}/config/schema:
    get:
      tags: [Plugins]
      summary: Get plugin configuration schema
      operationId: getPluginConfigSchema
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/{plugin_id}/config-files/{config_key}:
    get:
      tags: [Plugins]
      summary: List uploaded files for a plugin file configuration item
      operationId: listPluginConfigFiles
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
        - name: config_key
          in: path
          required: true
          description: URL-encoded configuration key path.
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Plugins]
      summary: Upload files for a plugin file configuration item
      operationId: uploadPluginConfigFiles
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
        - name: config_key
          in: path
          required: true
          description: URL-encoded configuration key path.
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/{plugin_id}/config-files:
    delete:
      tags: [Plugins]
      summary: Delete a plugin file configuration upload
      operationId: deletePluginConfigFile
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PluginConfigFileDeleteRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/plugins/{plugin_id}/readme:
    get:
      tags: [Plugins]
      summary: Get plugin README content
      operationId: getPluginReadme
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
      responses:
        "200":
          $ref: "#/components/responses/Text"
  /api/v1/plugins/{plugin_id}/changelog:
    get:
      tags: [Plugins]
      summary: Get plugin CHANGELOG content
      operationId: getPluginChangelog
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
      responses:
        "200":
          $ref: "#/components/responses/Text"
  /api/v1/plugins/{plugin_id}/reload:
    post:
      tags: [Plugins]
      summary: Reload a plugin
      operationId: reloadPlugin
      x-astrbot-scope: plugin
      parameters:
        - $ref: "#/components/parameters/PluginId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    AttachmentIdQuery:
      name: attachment_id
      in: query
      required: true
      schema:
        type: string
...
    PluginId:
      name: plugin_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    DynamicConfig:
      type: object
      additionalProperties: true
...
    FileUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
...
    PluginConfigFileDeleteRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
      additionalProperties: false
...